Clean up deprecated code.
