+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:flutter_todo/views/login.dart';
 import 'package:flutter_todo/views/register.dart';
 import 'package:flutter_todo/views/password_reset.dart';
 import 'package:flutter_todo/views/add_todo.dart';
-import 'package:flutter_todo/views/tasks.dart';
+import 'package:flutter_todo/views/todos.dart';
 
 void main() {
   runApp(
